requirements.txt
streamlit
python-dotenv
google-generativeai
youtube-transcript-api
python-docx
requests
cachetools>=5.0.0
